python/requirements.txt
mysql-connector-python
clickhouse-connect
click
